Avoid re-setting the same resolution
It turns out that this is what takes the majority of the time in
the process of grabbing a frame from the camera. Since most users
are likely to repeatedly use the same resolution this avoids an
expensive operation they don't need, but is still safe if they do
need it.