Implement valid coordinate conversion
I'm not sure how the previous implementation ever worked (it seems
to even fail trivial tests), so replace that with a simpler one.

While we're at it, introduce typed named tuples for the coordinate
systems so that we get useful representations.

This also adds tests: both unit tests for the conversion util and
integration tests that we get the right values for some example
images.