Correct labelling of the tecknet 25cm calibrations
At some point it looks like the labels got mixed up. Assuming that
the set of labels was correct, I believe that this applies the
labels to the right images. Certainly this now ensures that the
sequences of offsets at each distance is in the right order.

Unfortunately this also removes a number of images which were not
only mis-labeled, but were also duplicates of other (also
mis-labeled) images. I've left the related (now missing) entries
in the files.yaml as I'll try to back-fill them with the proper
images, if I can find them. If not, I'll remove them.
Either way this commit deliberately records the name fixes separately
from the restoration of the missing images.