Produce an error when loading a mismatching calibration
If we use a calibration for a resolution other than that which it
is intended, we'll get nonsense values. Ensure that that doesn't
happen by requiring that `load_camera_calibrations` is passed the
image resolution which will be used and then validating that
against the value in the calibration file.